claude_package/requirements.txt
mcp>=1.2.0
httpx
pyyaml
requests
